refactor(widgets): finish the catalog trait groups
