+++ crates/ui/src/widgets/mod.rs
@@ -1,14 +1,25 @@
-//! Small controls and content pieces, grouped as catalog traits. The page
-//! rhythm and section cards live in the [`Scaffolding`] trait; each group is imported
-//! on its own (`use bezel_ui::widgets::Scaffolding;` → `theme.group_box()`).
+//! Widgets, grouped as catalog traits on `Theme` — import the group, reach
+//! the component: `use bezel_ui::widgets::{Content, Controls, Layout,
+//! Scaffolding, Status};` → `theme.group_box()`, `theme.tab(..)`.
+//!
+//! What stays here is deliberately trait-shaped: state flags, pure math, and
+//! the hover refinements, none of which reads the theme as a receiver.
 
-use gpui::{SharedString, div, prelude::*, px};
+use gpui::{div, prelude::*, px};
 
-use bezel_theme::{Theme, ink};
+use bezel_theme::ink;
 
+mod content;
+mod controls;
+mod layout;
 mod scaffolding;
+mod status;
 
+pub use content::Content;
+pub use controls::{Controls, SliderDrag};
+pub use layout::{Layout, SplitDrag};
 pub use scaffolding::{OPTION_CARD_HEIGHT, OPTION_CARD_RADIUS, Scaffolding};
+pub use status::Status;
 
 /// What a control paints in the 1px border it keeps for
 /// [`crate::focus::focusable`]'s ring: nothing, until focus fills it.
@@ -17,338 +28,6 @@ pub use scaffolding::{OPTION_CARD_HEIGHT, OPTION_CARD_RADIUS, Scaffolding};
 /// appeared only on focus would shift the content under it by a pixel — a
 /// checkbox whose tick jumps as you tab onto it.
 pub(crate) const RING_SLOT: gpui::Hsla = gpui::transparent_black();
-
-/// Right-anchored badge pill: `rounded-full border px-2 py-0.5 text-[10.5px]`.
-pub fn badge(theme: &Theme, label: impl Into<SharedString>) -> gpui::Div {
-    div()
-        .flex_none()
-        .px(px(8.0))
-        .py(px(2.0))
-        .rounded_full()
-        .border_1()
-        .border_color(theme.border)
-        .text_size(px(10.5))
-        .text_color(theme.text_muted)
-        .child(label.into())
-}
-
-/// Emerald status pill (the Accounts "Active" badge:
-/// `bg-emerald-400/[0.12] text-emerald-300/90`).
-pub fn badge_active(theme: &Theme, label: impl Into<SharedString>) -> gpui::Div {
-    let emerald = theme.success;
-    let emerald_text = theme.success_muted; // emerald-300
-    div()
-        .flex_none()
-        .px(px(8.0))
-        .py(px(2.0))
-        .rounded_full()
-        .bg(emerald.opacity(0.12))
-        .text_size(px(10.5))
-        .text_color(emerald_text.opacity(0.9))
-        .child(label.into())
-}
-
-/// Display-only toggle switch (the reference branch-picker.tsx `Toggle`): an 18×32
-/// pill whose knob slides right and track flips white when on. State is owned
-/// by the parent row — the caller adds `.id(..)` and `.on_click(..)`.
-pub fn toggle(theme: &Theme, on: bool) -> gpui::Div {
-    div()
-        .flex_none()
-        .w(px(32.0))
-        .h(px(18.0))
-        .rounded_full()
-        .bg(if on { theme.text } else { ink(0.15) })
-        .border_1()
-        .border_color(RING_SLOT)
-        .relative()
-        .child(
-            // One less than the 2px inset it looks like: absolute insets resolve
-            // against the padding box, which the ring slot has already moved in
-            // by a pixel.
-            div()
-                .absolute()
-                .top(px(1.0))
-                .left(px(if on { 15.0 } else { 1.0 }))
-                .size(px(14.0))
-                .rounded_full()
-                .bg(if on { theme.on_solid } else { ink(0.7) }),
-        )
-}
-
-/// Display-only checkbox: a 16px rounded square that fills with the text tone
-/// and shows a check when on. State is the caller's; add `.id(..)`/`.on_click(..)`.
-pub fn checkbox(theme: &Theme, checked: bool) -> gpui::Div {
-    let mut box_ = div()
-        .flex_none()
-        .size(px(16.0))
-        .rounded(px(4.0))
-        .flex()
-        .items_center()
-        .justify_center();
-    box_ = if checked {
-        box_.border_1().border_color(RING_SLOT).bg(theme.text)
-    } else {
-        box_.border_1().border_color(ink(0.25)).bg(ink(0.03))
-    };
-    if checked {
-        box_.child(
-            crate::icons::icon(crate::icons::CHECK)
-                .size(px(11.0))
-                .text_color(theme.on_solid),
-        )
-    } else {
-        box_
-    }
-}
-
-/// Display-only radio button: a 16px ring with an inner dot when selected.
-/// Radios are a *set* — the caller owns which index is on.
-pub fn radio_button(theme: &Theme, selected: bool) -> gpui::Div {
-    div()
-        .flex_none()
-        .size(px(16.0))
-        .rounded_full()
-        .border_1()
-        .border_color(if selected { theme.text } else { ink(0.25) })
-        .bg(ink(0.03))
-        .flex()
-        .items_center()
-        .justify_center()
-        .when(selected, |ring| {
-            ring.child(div().size(px(8.0)).rounded_full().bg(theme.text))
-        })
-}
-
-/// Determinate progress bar. `fraction` is clamped to `0..=1`; the track keeps
-/// its full width so the row never reflows as the value moves.
-pub fn progress_bar(theme: &Theme, fraction: f32) -> gpui::Div {
-    let fraction = fraction.clamp(0.0, 1.0);
-    div()
-        .w_full()
-        .h(px(4.0))
-        .rounded_full()
-        .bg(ink(0.12))
-        .child(
-            div()
-                .h_full()
-                .w(gpui::relative(fraction))
-                .rounded_full()
-                .bg(theme.text),
-        )
-}
-
-/// The drag payload of a [`slider`], shipped from here for the same reason
-/// [`SplitDrag`] is: one type per gesture, so two sliders in one window do not
-/// answer each other's `on_drag_move`.
-pub struct SliderDrag;
-
-/// Display-only slider: filled track behind a knob at `fraction` (clamped to
-/// `0..=1`). Dragging is the caller's — it owns the value and the mouse
-/// handlers; this is the paint.
-///
-/// The element *is* the drag source, so the gesture is grab-anywhere-and-slide,
-/// and [`axis_fraction`] turns the pointer into the value:
-///
-/// ```ignore
-/// focus::focusable(&theme, &self.slider, slider(&theme, self.level))
-///     .id("slider")
-///     .on_drag(SliderDrag, |_, _, _, cx| cx.new(|_| gpui::Empty))
-///     .on_drag_move(cx.listener(|view, event: &DragMoveEvent<SliderDrag>, _, cx| {
-///         view.level = axis_fraction(event.event.position, event.bounds, Axis::Horizontal, 0.0);
-///         cx.notify();
-///     }))
-/// ```
-pub fn slider(theme: &Theme, fraction: f32) -> gpui::Div {
-    let fraction = fraction.clamp(0.0, 1.0);
-    div()
-        .w_full()
-        .h(px(16.0))
-        .border_1()
-        .border_color(RING_SLOT)
-        .rounded(px(4.0))
-        .flex()
-        .items_center()
-        .relative()
-        .child(
-            div()
-                .w_full()
-                .h(px(4.0))
-                .rounded_full()
-                .bg(ink(0.12))
-                .child(
-                    div()
-                        .h_full()
-                        .w(gpui::relative(fraction))
-                        .rounded_full()
-                        .bg(theme.text),
-                ),
-        )
-        .child(
-            // Inset by the knob's own width so it never overhangs the track.
-            div().absolute().left(gpui::relative(fraction)).child(
-                div()
-                    .size(px(14.0))
-                    .ml(px(-7.0))
-                    .rounded_full()
-                    .bg(theme.text),
-            ),
-        )
-}
-
-/// Circular avatar holding one or two initials — the fallback every avatar
-/// needs, and often the whole thing on a monochrome surface.
-pub fn avatar(theme: &Theme, initials: impl Into<SharedString>) -> gpui::Div {
-    div()
-        .flex_none()
-        .size(px(28.0))
-        .rounded_full()
-        .bg(ink(0.12))
-        .flex()
-        .items_center()
-        .justify_center()
-        .text_size(px(11.0))
-        .font_weight(gpui::FontWeight::MEDIUM)
-        .text_color(theme.text_muted)
-        .child(initials.into())
-}
-
-/// The closed face of a select: current value plus a chevron, shaped and
-/// toned like [`crate::input::TextField`] so a form of fields and selects
-/// reads as one system.
-///
-/// There is no `Select` component, deliberately — a select IS this trigger
-/// plus [`crate::popover::anchored_menu_below`] over [`crate::popover::menu_row`]s,
-/// and the caller already owns the open state and the selection. Wrapping
-/// that in a struct would buy an abstraction and cost the caller its control
-/// over both.
-pub fn select_trigger(theme: &Theme, label: impl Into<SharedString>, open: bool) -> gpui::Div {
-    div()
-        .flex()
-        .flex_row()
-        .items_center()
-        .justify_between()
-        .gap(px(8.0))
-        .px(px(10.0))
-        .py(px(7.0))
-        .rounded(px(Theme::BUTTON_RADIUS))
-        .bg(theme.input_bg)
-        .border_1()
-        .border_color(if open { theme.caret } else { theme.border })
-        .text_size(px(13.0))
-        .text_color(theme.text)
-        .cursor_pointer()
-        .child(div().min_w_0().truncate().child(label.into()))
-        .child(
-            crate::icons::icon(crate::icons::ALT_ARROW_DOWN)
-                .size(px(14.0))
-                .text_color(theme.text_muted),
-        )
-}
-
-/// Segmented control: one pill holding mutually exclusive choices, for when
-/// there are few enough that a [`select_trigger`] would be overkill.
-///
-/// `self_start` because a segmented control must hug its segments: dropped into
-/// a `flex_col`, flexbox's default `align-items: stretch` would otherwise blow
-/// it out to the column's full width.
-/// The segmented track's radius, and the inset its segments come in by. Two
-/// numbers read from both here and [`toggle_group_item`], so a segment cannot
-/// stop being concentric with the track it sits in.
-const TOGGLE_GROUP_RADIUS: f32 = 9.0;
-const TOGGLE_GROUP_PAD: f32 = 2.0;
-
-pub fn toggle_group(theme: &Theme) -> gpui::Div {
-    div()
-        .self_start()
-        .flex()
-        .flex_row()
-        .items_center()
-        .gap(px(TOGGLE_GROUP_PAD))
-        .p(px(TOGGLE_GROUP_PAD))
-        .rounded(px(TOGGLE_GROUP_RADIUS))
-        .bg(ink(0.06))
-        .border_1()
-        .border_color(theme.border)
-}
-
-/// One segment. The selected segment carries a raised plate; the rest are
-/// bare, so exactly one reads as pressed.
-pub fn toggle_group_item(
-    theme: &Theme,
-    label: impl Into<SharedString>,
-    selected: bool,
-) -> gpui::Div {
-    let mut item = div()
-        .px(px(10.0))
-        .py(px(4.0))
-        // Concentric with the track: 9 - 2 = 7.
-        .rounded(px(Theme::inset_radius(
-            TOGGLE_GROUP_RADIUS,
-            TOGGLE_GROUP_PAD,
-        )))
-        .border_1()
-        .border_color(RING_SLOT)
-        .text_size(px(12.5))
-        .cursor_pointer()
-        .child(label.into());
-    item = if selected {
-        item.bg(theme.surface_raised)
-            .font_weight(gpui::FontWeight::MEDIUM)
-            .text_color(theme.text)
-    } else {
-        item.text_color(theme.text_muted)
-    };
-    item
-}
-
-/// The disclosure chevron: right when collapsed, down when expanded.
-///
-/// Two assets rather than one rotated: gpui has no transform for `div`s at the
-/// pinned rev, and an SVG rotation would need a transform on the element.
-pub fn disclosure(theme: &Theme, expanded: bool) -> gpui::Svg {
-    crate::icons::icon(if expanded {
-        crate::icons::ALT_ARROW_DOWN
-    } else {
-        crate::icons::ALT_ARROW_RIGHT
-    })
-    .size(px(14.0))
-    .text_color(theme.text_muted)
-}
-
-/// Header row of a collapsible section: chevron plus title. The caller owns
-/// `expanded` and renders the body itself — a container that swallowed its
-/// children would have to re-implement layout for them. Hover is caller-owned
-/// (gpui panics on a second hover); the default wash is
-/// [`collapsible_header_hover`].
-pub fn collapsible_header(
-    theme: &Theme,
-    label: impl Into<SharedString>,
-    expanded: bool,
-) -> gpui::Div {
-    div()
-        .self_start()
-        .flex()
-        .flex_row()
-        .items_center()
-        .gap(px(6.0))
-        .px(px(4.0))
-        .py(px(5.0))
-        .rounded(px(Theme::CONTROL_RADIUS))
-        .cursor_pointer()
-        .child(disclosure(theme, expanded))
-        .child(
-            div()
-                .text_size(px(12.5))
-                .font_weight(gpui::FontWeight::MEDIUM)
-                .text_color(theme.text)
-                .child(label.into()),
-        )
-}
-
-/// The default collapsible-header hover wash (`hover:bg-white/[0.03]`).
-pub fn collapsible_header_hover(s: gpui::StyleRefinement) -> gpui::StyleRefinement {
-    s.bg(ink(0.03))
-}
 
 /// A flag that follows something else until the user takes it over.
 ///
@@ -384,151 +63,6 @@ impl Takeover {
     }
 }
 
-/// Padding inside a [`step_row`] and the [`step_output`] under it — the two
-/// have to agree or the output's first character sits left of the title.
-const STEP_PAD_X: f32 = 10.0;
-const STEP_PAD_Y: f32 = 6.0;
-/// How tall an output may get before it scrolls (`ToolCard.svelte`'s
-/// `max-h-64`).
-const STEP_OUTPUT_MAX: f32 = 256.0;
-
-/// One operation, as a row: an icon, what it was, and how it went.
-///
-/// A tool call in a transcript, a step in a CI run, a file in a migration — the
-/// shape is the same everywhere, which is why this takes strings and not a type
-/// that knows what any of them mean. `detail` is the truncating middle (a
-/// query, a path, a `· 3` count), `meta` the right-aligned figure that never
-/// truncates (a duration, a size, a row count).
-///
-/// `expanded` is `None` when there is nothing under the row, and the chevron is
-/// simply absent — a disclosure that opens onto nothing is worse than no
-/// disclosure. `Some` renders it, and the caller owns the flag.
-///
-/// Returns a plain `Div` like the rest of this module: the caller adds `.id(..)`
-/// and `.on_click(..)` **to this row**, never to a wrapper around it, or the
-/// hitbox ends up narrower than what it paints. Hover is caller-owned (gpui
-/// panics on a second hover); the default wash is [`step_row_hover`].
-pub fn step_row(
-    theme: &Theme,
-    icon: &'static str,
-    title: impl Into<SharedString>,
-    detail: Option<SharedString>,
-    meta: Option<SharedString>,
-    failed: bool,
-    expanded: Option<bool>,
-) -> gpui::Div {
-    div()
-        .flex()
-        .flex_row()
-        .items_center()
-        .gap(px(8.0))
-        .px(px(STEP_PAD_X))
-        .py(px(STEP_PAD_Y))
-        .cursor_pointer()
-        .child(
-            // Tinted here rather than on the row: gpui reads an svg's colour
-            // off its own style, so a colour set on the parent never arrives.
-            crate::icons::icon(icon)
-                .size(px(14.0))
-                .text_color(if failed {
-                    theme.danger
-                } else {
-                    theme.text_muted
-                }),
-        )
-        .child(
-            div()
-                .flex_none()
-                .text_size(px(12.5))
-                .text_color(theme.text)
-                .child(title.into()),
-        )
-        .when_some(detail, |row, detail| {
-            row.child(
-                div()
-                    .min_w_0()
-                    .truncate()
-                    .text_size(px(12.0))
-                    .text_color(theme.text_muted)
-                    .child(detail),
-            )
-        })
-        .child(
-            div()
-                .ml_auto()
-                .flex_none()
-                .flex()
-                .flex_row()
-                .items_center()
-                .gap(px(6.0))
-                .when_some(meta, |cluster, meta| {
-                    cluster.child(
-                        div()
-                            .font_family(theme.font_mono.clone())
-                            .text_size(px(12.0))
-                            .text_color(theme.text_muted)
-                            .child(meta),
-                    )
-                })
-                .when_some(expanded, |cluster, expanded| {
-                    cluster.child(disclosure(theme, expanded))
-                }),
-        )
-}
-
-/// The default card-row hover wash (`hover:bg-white/[0.015]`).
-pub fn card_row_hover(s: gpui::StyleRefinement) -> gpui::StyleRefinement {
-    s.bg(ink(0.015))
-}
-
-/// The default step-row hover wash (`hover:bg-white/[0.03]`).
-pub fn step_row_hover(s: gpui::StyleRefinement) -> gpui::StyleRefinement {
-    s.bg(ink(0.03))
-}
-
-/// What a [`step_row`] opens onto: its output, verbatim.
-///
-/// Monospaced and capped, because the thing being shown is a program's stdout
-/// and the row it hangs off is one line tall — a 900-line stack trace pushing
-/// the next step off screen is the failure this cap exists for. Past the cap it
-/// scrolls, which is why it takes an id.
-///
-/// No scrollbar: the wheel reaches it regardless, and a bar would need a
-/// `ScrollHandle` and a `ScrollbarState` from every caller for a box that is
-/// usually four lines long. Wrap it in `div().relative()` with
-/// [`crate::scroll::scrollbar`] over it if a particular one earns the bar.
-pub fn step_output(
-    theme: &Theme,
-    id: impl Into<gpui::ElementId>,
-    text: impl Into<SharedString>,
-) -> gpui::Stateful<gpui::Div> {
-    div()
-        .id(id)
-        .max_h(px(STEP_OUTPUT_MAX))
-        .overflow_y_scroll()
-        .border_t_1()
-        .border_color(theme.border)
-        .px(px(STEP_PAD_X))
-        .py(px(STEP_PAD_Y))
-        .font_family(theme.font_mono.clone())
-        .text_size(px(12.0))
-        .text_color(theme.text_muted)
-        .child(text.into())
-}
-
-// ---------------------------------------------------------------------------
-// Resizable split
-// ---------------------------------------------------------------------------
-
-/// The drag payload of a [`split_handle`]. Shipped from here so every split
-/// speaks the same type: `on_drag_move::<SplitDrag>` on one container would
-/// otherwise fire for an unrelated split's gesture.
-pub struct SplitDrag;
-
-/// Width of the divider's grab strip: the 1px line plus zed's own 4px of slack
-/// each side (`workspace::HANDLE_HITBOX_SIZE`) — a 1px target is unhittable.
-const SPLIT_HANDLE_HIT: f32 = 9.0;
-
 /// Where `pointer` falls along `axis` as a fraction of `bounds` — what a
 /// divider dragged there makes the split, and what a slider dragged there makes
 /// the value. `Axis::Horizontal` travels in x.
@@ -554,272 +88,23 @@ pub fn axis_fraction(
     (offset / extent).clamp(min, 1.0 - min)
 }
 
-/// The divider between two panes: a hairline centred in a grab strip, lit while
-/// dragged.
-///
-/// The gesture stays with the caller, which owns the fraction — the handle is
-/// dragged with gpui's null-preview drag, and the container reads the pointer:
-///
-/// ```ignore
-/// div()
-///     .id("split")
-///     .on_drag_move(cx.listener(|view, event: &DragMoveEvent<SplitDrag>, _, cx| {
-///         view.fraction = axis_fraction(event.event.position, event.bounds, Axis::Horizontal, 0.15);
-///         cx.notify();
-///     }))
-///     .child(div().w(relative(self.fraction)).child(left))
-///     .child(
-///         split_handle(&theme, Axis::Horizontal, self.dragging)
-///             .id("split-handle")
-///             .on_drag(SplitDrag, |_, _, _, cx| cx.new(|_| gpui::Empty)),
-///     )
-///     .child(div().flex_1().child(right))
-/// ```
-pub fn split_handle(theme: &Theme, axis: gpui::Axis, dragging: bool) -> gpui::Div {
-    let line = if dragging { theme.caret } else { theme.border };
-    let handle = div().flex_none().flex().items_center().justify_center();
-    match axis {
-        gpui::Axis::Horizontal => handle
-            .w(px(SPLIT_HANDLE_HIT))
-            .h_full()
-            .cursor_col_resize()
-            .child(div().w(px(1.0)).h_full().bg(line)),
-        gpui::Axis::Vertical => handle
-            .h(px(SPLIT_HANDLE_HIT))
-            .w_full()
-            .cursor_row_resize()
-            .child(div().h(px(1.0)).w_full().bg(line)),
-    }
-}
-
-/// A removable chip — a token in a filter bar or a recipient field. The caller
-/// adds the click handler for the ✕.
-pub fn tag(theme: &Theme, label: impl Into<SharedString>) -> gpui::Div {
-    div()
-        .self_start()
-        .flex()
-        .flex_row()
-        .items_center()
-        .gap(px(5.0))
-        .pl(px(8.0))
-        .pr(px(5.0))
-        .py(px(3.0))
-        .rounded(px(Theme::CONTROL_RADIUS))
-        .bg(ink(0.07))
-        .border_1()
-        .border_color(theme.border)
-        .text_size(px(12.0))
-        .text_color(theme.text)
-        .child(label.into())
-        .child(
-            crate::icons::icon(crate::icons::CLOSE)
-                .size(px(10.0))
-                .text_color(theme.text_faint),
-        )
-}
-
-/// Breadcrumb trail. Items are added by the caller with [`breadcrumb_item`],
-/// separated by [`breadcrumb_separator`].
-pub fn breadcrumb() -> gpui::Div {
-    div()
-        .self_start()
-        .flex()
-        .flex_row()
-        .items_center()
-        .gap(px(4.0))
-        .min_w_0()
-}
-
-/// One crumb. The last one is `current` and stops looking clickable.
-pub fn breadcrumb_item(theme: &Theme, label: impl Into<SharedString>, current: bool) -> gpui::Div {
-    let mut crumb = div()
-        .min_w_0()
-        .truncate()
-        .text_size(px(12.5))
-        .child(label.into());
-    crumb = if current {
-        crumb.text_color(theme.text)
-    } else {
-        crumb.text_color(theme.text_muted).cursor_pointer()
-    };
-    crumb
-}
-
-/// The chevron between crumbs.
-pub fn breadcrumb_separator(theme: &Theme) -> gpui::Svg {
-    crate::icons::icon(crate::icons::ALT_ARROW_RIGHT)
-        .size(px(12.0))
-        .text_color(theme.text_faint)
-}
-
 /// A small state dot — the "working / idle / failed" bead on a row. Takes the
 /// tone from the caller so the meaning stays with the caller's domain.
 pub fn status_dot(tone: gpui::Hsla) -> gpui::Div {
     div().flex_none().size(px(6.0)).rounded_full().bg(tone)
 }
 
-/// The centered "nothing here yet" panel: icon, headline, one line of hint.
-pub fn empty_state(
-    theme: &Theme,
-    icon_path: &'static str,
-    title: impl Into<SharedString>,
-    hint: impl Into<SharedString>,
-) -> gpui::Div {
-    div()
-        .w_full()
-        .flex()
-        .flex_col()
-        .items_center()
-        .justify_center()
-        .gap(px(6.0))
-        .py(px(40.0))
-        .child(
-            crate::icons::icon(icon_path)
-                .size(px(24.0))
-                .text_color(theme.text_faint),
-        )
-        .child(
-            div()
-                .text_size(px(13.5))
-                .font_weight(gpui::FontWeight::MEDIUM)
-                .text_color(theme.text)
-                .child(title.into()),
-        )
-        .child(
-            div()
-                .text_size(px(12.5))
-                .text_color(theme.text_muted)
-                .child(hint.into()),
-        )
+/// The default card-row hover wash (`hover:bg-white/[0.015]`).
+pub fn card_row_hover(s: gpui::StyleRefinement) -> gpui::StyleRefinement {
+    s.bg(ink(0.015))
 }
 
-/// Tab strip: a hairline-underlined row that tabs sit on.
-pub fn tab_bar(theme: &Theme) -> gpui::Div {
-    div()
-        .flex()
-        .flex_row()
-        .items_center()
-        .gap(px(2.0))
-        .border_b_1()
-        .border_color(theme.border)
+/// The default step-row hover wash (`hover:bg-white/[0.03]`).
+pub fn step_row_hover(s: gpui::StyleRefinement) -> gpui::StyleRefinement {
+    s.bg(ink(0.03))
 }
 
-/// One tab. The active tab is marked by the text tone plus a 2px underline that
-/// overlaps the bar's hairline, so switching tabs never changes row height.
-pub fn tab(theme: &Theme, label: impl Into<SharedString>, active: bool) -> gpui::Div {
-    div()
-        .relative()
-        .px(px(10.0))
-        .pb(px(7.0))
-        .pt(px(6.0))
-        .rounded_t(px(Theme::CONTROL_RADIUS))
-        .border_1()
-        .border_color(RING_SLOT)
-        .text_size(px(13.0))
-        .font_weight(if active {
-            gpui::FontWeight::MEDIUM
-        } else {
-            gpui::FontWeight::NORMAL
-        })
-        .text_color(if active { theme.text } else { theme.text_muted })
-        .cursor_pointer()
-        .child(label.into())
-        .when(active, |t| {
-            t.child(
-                // Insets resolve against the padding box, so each one carries
-                // the ring slot's pixel: the underline still spans the tab's
-                // full width and still overlaps the bar's hairline.
-                div()
-                    .absolute()
-                    .bottom(px(-2.0))
-                    .left(px(-1.0))
-                    .right(px(-1.0))
-                    .h(px(2.0))
-                    .bg(theme.text),
-            )
-        })
-}
-
-/// A small quiet ghost action (`rounded-lg px-2.5 py-1.5 text-[12px]
-/// text-muted-foreground`). Caller adds id + click + leading icon child AND
-/// its own `.hover(..)` — gpui panics on a second hover, and the pages vary
-/// it (reveal opacity, 4% vs 6% washes).
-pub fn ghost_action(theme: &Theme) -> gpui::Div {
-    div()
-        .flex()
-        .flex_row()
-        .items_center()
-        .gap(px(6.0))
-        .rounded(px(Theme::BUTTON_RADIUS))
-        .px(px(10.0))
-        .py(px(6.0))
-        .text_size(px(12.0))
-        .text_color(theme.text_muted)
-        .cursor_pointer()
-}
-
-/// The default ghost-action hover wash (`hover:bg-white/[0.06]
-/// hover:text-foreground`).
-pub fn ghost_hover(theme: &Theme, s: gpui::StyleRefinement) -> gpui::StyleRefinement {
-    s.bg(ink(0.06)).text_color(theme.text)
-}
-
-/// The dismissible red error strip (`flex items-start gap-2 rounded-xl border
-/// border-red-400/20 bg-red-400/[0.06] text-red-300/90` with a leading
-/// `DangerTriangle mt-0.5 size-4`).
-pub fn error_strip(theme: &Theme, message: impl Into<SharedString>) -> gpui::Div {
-    let red = theme.danger; // red-400
-    let red_text = theme.danger_muted; // red-300
-    div()
-        .mt(px(16.0))
-        .px(px(16.0))
-        .py(px(12.0))
-        .rounded(px(Theme::SURFACE_RADIUS))
-        .border_1()
-        .border_color(red.opacity(0.2))
-        .bg(red.opacity(0.06))
-        .text_size(px(12.5))
-        .text_color(red_text.opacity(0.9))
-        .flex()
-        .flex_row()
-        .items_start()
-        .gap(px(8.0))
-        .child(
-            div().flex_none().mt(px(2.0)).child(
-                crate::icons::icon(crate::icons::DANGER_TRIANGLE)
-                    .size(px(16.0))
-                    .text_color(red_text.opacity(0.9)),
-            ),
-        )
-        .child(div().min_w_0().child(message.into()))
-}
-
-/// The amber warning strip (`flex items-start gap-2 border-amber-400/20
-/// bg-amber-400/[0.06] text-amber-200/90` with a leading `DangerTriangle
-/// mt-0.5 size-3.5`).
-pub fn warning_strip(theme: &Theme, message: impl Into<SharedString>) -> gpui::Div {
-    let amber = theme.warning; // amber-400
-    let amber_text = theme.warning_muted; // amber-200
-    div()
-        .mt(px(8.0))
-        .px(px(16.0))
-        .py(px(10.0))
-        .rounded(px(Theme::SURFACE_RADIUS))
-        .border_1()
-        .border_color(amber.opacity(0.2))
-        .bg(amber.opacity(0.06))
-        .text_size(px(12.0))
-        .text_color(amber_text.opacity(0.9))
-        .flex()
-        .flex_row()
-        .items_start()
-        .gap(px(8.0))
-        .child(
-            div().flex_none().mt(px(2.0)).child(
-                crate::icons::icon(crate::icons::DANGER_TRIANGLE)
-                    .size(px(14.0))
-                    .text_color(amber_text.opacity(0.9)),
-            ),
-        )
-        .child(div().min_w_0().child(message.into()))
+/// The default collapsible-header hover wash (`hover:bg-white/[0.03]`).
+pub fn collapsible_header_hover(s: gpui::StyleRefinement) -> gpui::StyleRefinement {
+    s.bg(ink(0.03))
 }
